Handle 400 error from code EP
