+++ client/landing/stepper/hooks/index.d.ts
@@ -1,0 +1,7 @@
+type SHA256 = {
+	update( data: string ): SHA256;
+	digest( encoding: 'hex' ): string;
+};
+declare module 'hash.js/lib/hash/sha/256' {
+	export default function sha256(): SHA256;
+}
