+++ packages/data-stores/src/index.ts
@@ -31,7 +31,6 @@ export * from './domain-suggestions/types';
 export * from './plans/types';
 export * from './user/types';
 export * from './queries/use-launchpad';
-export * from './queries/use-is-domain-code-valid';
 
 const { SubscriptionManager } = Reader;
